Move get_*_name to StrawberryConfig
